🧪 test(web): isolate i18n coverage
Keep router tests focused on query-client behavior and move locale coverage to its own suite.
